Faster at refreshing data when we activated something.
Calling the relevant Charge, Climate or Vehicle state in their own handler for features that are visible on screen instead of relying on a more time/bandwidth intensive getVehicleData. Features  handled here are lock/unlock, opening/closing something, turning the climate on/off as well as sentry and scheduled departing.

vehicle data samples are timestamped and if a sample is older than the current one, it is ignored. Having two timers possibly running at the same time could create this scenario.

delayedRetry was simply calling stateMachine now so timers calls stateMachine directly now. delayedRetry was retired.

When we're waiting for refreshed data from activating something, the main stateMachine is 'paused' to prevent 'colliding' with our call have obsolete data.